Started to implement the model
Code is a bit of a mess. I think conv lstm is expecting batch, time step, feature, height, width, Even for the inputs so there are a lot of reshaping as I was trying to make it work. Also have a look at how I initialized the model, it might be wrong.