Tried to modify train_epoch
I tried to use the previous outputs as the current inputs, but I think I am missing something. I am not sure if the class does this for us already or not, which is why I am confused.
Val_epoch is not yet implemented, only work on it once train_epoch is debugged