Saved scalers in a new subfolder and added thresholds for metrics
Scalers are now saved during the training notebook to make them easier to call in the testing notebooks.
Thresholds were added to the metrics to take into account small water depths predicted by the model. A large threshold decreases recall as there is now a larger number of negatives, creating a disproportion in the confusion matrix.
Still need to update docstrings